+++ examples/concertbot/train.py
@@ -11,13 +11,12 @@ async def train_dialogue(domain_file='domain.yml',
                          stories_file='data/stories.md',
                          model_path='models/dialogue',
                          policy_config='policy_config.yml'):
-    return await train.train_dialogue_model(domain_file=domain_file,
-                                            stories_file=stories_file,
-                                            output_path=model_path,
-                                            policy_config=policy_config,
-                                            kwargs={'augmentation_factor': 50,
-                                                    'validation_split': 0.2}
-                                            )
+    return await train(domain_file=domain_file,
+                       stories_file=stories_file,
+                       output_path=model_path,
+                       policy_config=policy_config,
+                       kwargs={'augmentation_factor': 50,
+                               'validation_split': 0.2})
 
 
 if __name__ == '__main__':