+++ rasa_core/nlg/template.py
@@ -1,7 +1,7 @@
 import copy
 import logging
+import re
 
-import numpy as np
 from rasa_core.trackers import DialogueStateTracker
 from typing import Text, Any, Dict, Optional, List
 
@@ -25,6 +25,7 @@ class TemplatedNaturalLanguageGenerator(NaturalLanguageGenerator):
                              output_channel: Text
                              ) -> Optional[Dict[Text, Any]]:
         """Select random template for the utter action from available ones."""
+        import numpy as np
 
         if utter_action in self.templates:
             return np.random.choice(self.templates[utter_action])
@@ -76,7 +77,13 @@ class TemplatedNaturalLanguageGenerator(NaturalLanguageGenerator):
         # Filling the template variables in the template
         if template_vars:
             try:
-                template["text"] = template["text"].format(**template_vars)
+                # transforming template tags from
+                # "{tag_name}" to "{0[tag_name]}"
+                # as described here:
+                # https://stackoverflow.com/questions/7934620/python-dots-in-the-name-of-variable-in-a-format-string#comment9695339_7934969
+                # assuming that slot_name do not contain newline character here
+                text = re.sub(r'{([^\n]+?)}', r'{0[\1]}', template["text"])
+                template["text"] = text.format(template_vars)
             except KeyError as e:
                 logger.exception(
                     "Failed to fill utterance template '{}'. "