+++ rasa_core/utils.py
@@ -1,70 +1,43 @@
 # -*- coding: utf-8 -*-
-import errno
-import sys
-
-import aiohttp
 import argparse
 import asyncio
-import inspect
+import errno
 import io
 import json
 import logging
 import os
 import re
+import sys
+import tarfile
 import tempfile
 import warnings
+import zipfile
+from asyncio import AbstractEventLoop
+from hashlib import md5, sha1
+from io import BytesIO as IOReader, StringIO
+from typing import Any, Dict, List, Optional, Set, TYPE_CHECKING, Text, Tuple
+
+import aiohttp
 from aiohttp import InvalidURL
-from asyncio import AbstractEventLoop, Future
-from hashlib import sha1
-from io import StringIO
-from numpy import all, array
-from random import Random
+from requests.exceptions import InvalidURL
 from sanic import Sanic
 from sanic.request import Request
 from sanic.views import CompositionView
-from typing import Text, Any, List, Optional, Tuple, Dict, Set
 
 from rasa_core.constants import DEFAULT_REQUEST_TIMEOUT
-from rasa_nlu import utils as nlu_utils
 
 logger = logging.getLogger(__name__)
+
+if TYPE_CHECKING:
+    from random import Random
 
 
 def configure_file_logging(loglevel, logfile):
     if logfile:
-        fh = logging.FileHandler(logfile)
+        fh = logging.FileHandler(logfile, encoding='utf-8')
         fh.setLevel(loglevel)
         logging.getLogger('').addHandler(fh)
     logging.captureWarnings(True)
-
-
-def add_logging_option_arguments(parser):
-    """Add options to an argument parser to configure logging levels."""
-
-    # arguments for logging configuration
-    parser.add_argument(
-        '-v', '--verbose',
-        help="Be verbose. Sets logging level to INFO",
-        action="store_const",
-        dest="loglevel",
-        const=logging.INFO,
-        default=logging.INFO,
-    )
-    parser.add_argument(
-        '-vv', '--debug',
-        help="Print lots of debugging statements. "
-             "Sets logging level to DEBUG",
-        action="store_const",
-        dest="loglevel",
-        const=logging.DEBUG,
-    )
-    parser.add_argument(
-        '--quiet',
-        help="Be quiet! Sets logging level to WARNING",
-        action="store_const",
-        dest="loglevel",
-        const=logging.WARNING,
-    )
 
 
 # noinspection PyUnresolvedReferences
@@ -126,7 +99,7 @@ def dump_obj_as_str_to_file(filename: Text, text: Text) -> None:
 def subsample_array(arr: List[Any],
                     max_values: int,
                     can_modify_incoming_array: bool = True,
-                    rand: Optional[Random] = None) -> List[Any]:
+                    rand: Optional['Random'] = None) -> List[Any]:
     """Shuffles the array and returns `max_values` number of elements."""
     import random
 
@@ -247,12 +220,24 @@ class bcolors(object):
     UNDERLINE = '\033[4m'
 
 
-def wrap_with_color(text, color):
+def wrap_with_color(text: Text, color: Text):
     return color + text + bcolors.ENDC
 
 
-def print_color(text, color):
+def print_color(text: Text, color: Text):
     print(wrap_with_color(text, color))
+
+
+def print_warning(text: Text):
+    print_color(text, bcolors.WARNING)
+
+
+def print_error(text: Text):
+    print_color(text, bcolors.FAIL)
+
+
+def print_success(text: Text):
+    print_color(text, bcolors.OKGREEN)
 
 
 class HashableNDArray(object):
@@ -279,11 +264,15 @@ class HashableNDArray(object):
             Optional. If True, a copy of the input ndaray is created.
             Defaults to False.
         """
+        from numpy import array
+
         self.__tight = tight
         self.__wrapped = array(wrapped) if tight else wrapped
         self.__hash = int(sha1(wrapped.view()).hexdigest(), 16)
 
     def __eq__(self, other):
+        from numpy import all
+
         return all(self.__wrapped == other.__wrapped)
 
     def __hash__(self):
@@ -294,6 +283,7 @@ class HashableNDArray(object):
 
         If the wrapper is "tight", a copy of the encapsulated ndarray is
         returned. Otherwise, the encapsulated ndarray itself is returned."""
+        from numpy import array
 
         if self.__tight:
             return array(self.__wrapped)
@@ -326,12 +316,12 @@ def replace_environment_variables():
     yaml.SafeConstructor.add_constructor(u'!env_var', env_var_constructor)
 
 
-def read_yaml_file(filename):
+def read_yaml_file(filename: Text) -> Dict[Text, Any]:
     """Read contents of `filename` interpreting them as yaml."""
     return read_yaml_string(read_file(filename))
 
 
-def read_yaml_string(string):
+def read_yaml_string(string: Text) -> Dict[Text, Any]:
     replace_environment_variables()
     import ruamel.yaml
 
@@ -339,7 +329,7 @@ def read_yaml_string(string):
     yaml_parser.version = "1.1"
     yaml_parser.unicode_supplementary = True
 
-    return yaml_parser.load(string)
+    return yaml_parser.load(string) or {}
 
 
 def _dump_yaml(obj, output):
@@ -427,6 +417,23 @@ def zip_folder(folder):
     return shutil.make_archive(zipped_path.name, str("zip"), folder)
 
 
+def unarchive(byte_array: bytes, directory: Text) -> Text:
+    """Tries to unpack a byte array interpreting it as an archive.
+
+    Tries to use tar first to unpack, if that fails, zip will be used."""
+
+    try:
+        tar = tarfile.open(fileobj=IOReader(byte_array))
+        tar.extractall(directory)
+        tar.close()
+        return directory
+    except tarfile.TarError:
+        zip_ref = zipfile.ZipFile(IOReader(byte_array))
+        zip_ref.extractall(directory)
+        zip_ref.close()
+        return directory
+
+
 def cap_length(s, char_limit=20, append_ellipsis=True):
     """Makes sure the string doesn't exceed the passed char limit.
 
@@ -504,6 +511,7 @@ def extract_args(kwargs: Dict[Text, Any],
 
 def arguments_of(func):
     """Return the parameters of the function `func` as a list of names."""
+    import inspect
 
     return list(inspect.signature(func).parameters.keys())
 
@@ -569,11 +577,28 @@ def read_lines(filename, max_line_limit=None, line_pattern=".*"):
                 break
 
 
+def file_as_bytes(path: Text) -> bytes:
+    """Read in a file as a byte array."""
+    with io.open(path, 'rb') as f:
+        return f.read()
+
+
+def get_file_hash(path: Text) -> Text:
+    """Calculate the md5 hash of a file."""
+    return md5(file_as_bytes(path)).hexdigest()
+
+
+def get_text_hash(text: Text, encoding: Text = "utf-8") -> Text:
+    """Calculate the md5 hash of a file."""
+    return md5(text.encode(encoding)).hexdigest()
+
+
 async def download_file_from_url(url: Text) -> Text:
     """Download a story file from a url and persists it into a temp file.
 
     Returns the file path of the temp file that contains the
     downloaded content."""
+    from rasa_nlu import utils as nlu_utils
 
     if not nlu_utils.is_url(url):
         raise InvalidURL(url)
@@ -642,7 +667,7 @@ class ClientResponseError(aiohttp.ClientResponseError):
 class EndpointConfig(object):
     """Configuration for an external HTTP endpoint."""
 
-    def __init__(self, url, params=None, headers=None, basic_auth=None,
+    def __init__(self, url=None, params=None, headers=None, basic_auth=None,
                  token=None, token_name="token", **kwargs):
         self.url = url
         self.params = params if params else {}
@@ -650,7 +675,7 @@ class EndpointConfig(object):
         self.basic_auth = basic_auth
         self.token = token
         self.token_name = token_name
-        self.store_type = kwargs.pop('store_type', None)
+        self.type = kwargs.pop('store_type', kwargs.pop('type', None))
         self.kwargs = kwargs
         self._session = None