+++ tests/test_evaluation.py
@@ -4,8 +4,7 @@ import pytest
 # noinspection PyUnresolvedReferences
 from sklearn.exceptions import UndefinedMetricWarning
 
-from rasa_core import evaluate
-from rasa_core.evaluate import (collect_story_predictions, run_story_evaluation)
+from rasa_core.test import (test, collect_story_predictions, _generate_trackers)
 from tests.conftest import (
     DEFAULT_STORIES_FILE, E2E_STORY_FILE_UNKNOWN_ENTITY, END_TO_END_STORY_FILE)
 
@@ -17,8 +16,8 @@ async def test_evaluation_image_creation(tmpdir, default_agent):
     stories_path = os.path.join(tmpdir.strpath, "failed_stories.md")
     img_path = os.path.join(tmpdir.strpath, "story_confmat.pdf")
 
-    await run_story_evaluation(
-        resource_name=DEFAULT_STORIES_FILE,
+    await test(
+        stories=DEFAULT_STORIES_FILE,
         agent=default_agent,
         out_directory=tmpdir.strpath,
         max_stories=None,
@@ -30,7 +29,7 @@ async def test_evaluation_image_creation(tmpdir, default_agent):
 
 
 async def test_action_evaluation_script(tmpdir, default_agent):
-    completed_trackers = await evaluate._generate_trackers(
+    completed_trackers = await _generate_trackers(
         DEFAULT_STORIES_FILE, default_agent, use_e2e=False)
     story_evaluation, num_stories = collect_story_predictions(
         completed_trackers,
@@ -44,7 +43,7 @@ async def test_action_evaluation_script(tmpdir, default_agent):
 
 
 async def test_end_to_end_evaluation_script(tmpdir, default_agent):
-    completed_trackers = await evaluate._generate_trackers(
+    completed_trackers = await _generate_trackers(
         END_TO_END_STORY_FILE, default_agent, use_e2e=True)
 
     story_evaluation, num_stories = collect_story_predictions(
@@ -61,7 +60,7 @@ async def test_end_to_end_evaluation_script(tmpdir, default_agent):
 @pytest.mark.filterwarnings("ignore::UndefinedMetricWarning")
 async def test_end_to_end_evaluation_script_unknown_entity(tmpdir,
                                                            default_agent):
-    completed_trackers = await evaluate._generate_trackers(
+    completed_trackers = await _generate_trackers(
         E2E_STORY_FILE_UNKNOWN_ENTITY, default_agent, use_e2e=True)
 
     story_evaluation, num_stories = collect_story_predictions(